backend/app/schemas/resume_upload_schema.py
from pydantic import BaseModel, Field
from typing import Optional, List, Dict
from datetime import datetime

# Resume Upload Schema

class ResumeUploadResponse(BaseModel):
    candidate_id: str
    status: str
    message: str
    
    