backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.3.2
pymongo==4.6.1
redis[asyncio]==5.0.6
pydantic-settings==2.3.1
openai==2.15.0
pdfplumber==0.11.1
python-docx==1.1.2
httpx==0.27.0
python-multipart==0.0.9
reportlab==4.2.2
